nlp/tools/langchain-file-processor/requirements.txt
streamlit
langchain
langchain-chroma
langchain-google-genai
python-dotenv
chromadb
pypdf
